manage puppet agent version with the puppet_agent module
"auto" means match the version used on the server. Collection "puppet"
is the rolling release.

Also turned off unattended upgrades for Puppet packages, since that
conflicts with the purpose of the puppet_agent module. All I should
have to do now is upgrade the puppet server, then all the agents will
be upgraded.