puppet_runtime munin plugin: improve reporting reliability
Apparently the "--lines" option of journalctl takes precedence over
the "--grep" option, because when puppet is in the middle of a run,
the previous journalctl command wouldn't output anything. So now we
look at the last 1000 lines that contain "Applied catalog" and take
the last line with tail. In case that still fails, output U so that
munin doesn't get invalid input.